completed trader middleware
